# swagger.yaml
swagger: '2.0'
info:
  contact: {}
  title: ITMO Chan
  version: 1.0.0
basePath: /api
paths:
...
  /poll/18:
    post:
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          schema:
            example:
              - 2
            items:
              example: 2
              type: number
            type: array
      responses:
        '200':
          description: ''
      tags:
        - Poll
      description: Vote in poll
      operationId: voteInPoll
      summary: Vote in poll
...
  /thread:
    post:
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: comment
          type: string
        - format: binary
          in: formData
          name: files
          type: string
        - in: formData
          name: poll
          type: string
        - in: formData
          name: thread
          type: string
      responses:
        '200':
          description: ''
      tags:
        - Thread
      description: Add thread
      operationId: addThread
      summary: Add thread
...
  /topic:
    get:
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Topic
      description: Get all topics
      operationId: getAllTopics
      summary: Get all topics
    put:
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          schema:
            properties:
              description:
                example: Mega vazhno
                type: string
              name:
                example: topic432
                type: string
            type: object
      responses:
        '200':
          description: ''
      tags:
        - Topic
      description: Add topic
      operationId: addTopic
      summary: Add topic
...